requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.35.0
torch==2.1.0
openai==1.3.0
tenacity==8.2.3
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
scikit-learn==1.3.2
google-generativeai==0.3.1
